# web/docs/alarm_rules_manager/alarm_v1_swagger.yaml
basePath: /api/v1/rules
definitions:
...
  apis.SecruleListResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      security_rules:
        items:
          $ref: '#/definitions/apis.SecruleResponse'
        type: array
      total:
        type: integer
    type: object
  apis.SecruleResponse:
    properties:
      created_at:
        type: string
      direction:
        type: string
      id:
        type: string
      ip_version:
        type: string
      name:
        type: string
      owner:
        type: string
      port_max:
        type: integer
      port_min:
        type: integer
      protocol:
        type: string
      remote_cidr:
        type: string
      remote_group:
        $ref: '#/definitions/common.BaseReference'
      updated_at:
        type: string
    type: object
...
  apis.SecurityRulePayload:
    properties:
      direction:
        enum:
        - ingress
        - egress
        type: string
      port_max:
        maximum: 65535
        minimum: 1
        type: integer
      port_min:
        maximum: 65535
        minimum: 1
        type: integer
      protocol:
        enum:
        - tcp
        - udp
        - icmp
        type: string
      remote_cidr:
        type: string
    required:
    - direction
    - protocol
    type: object
...
  common.APIError:
    properties:
      error_message:
        description: "InternalErr error\n\tErrorCode int `json:\"error_code\"`"
        type: string
    type: object
...
  common.BaseReference:
    properties:
      id:
        type: string
      name:
        maxLength: 32
        minLength: 2
        type: string
    type: object
...
  main.RuleFileRequest:
    description: Request for rule file operations
    properties:
      content:
        description: |-
          File content for write operation
          @Description Content to write to file (required for write operation)
        type: string
      file_path:
        description: |-
          Source file path
          @Description Path of the source file (required for all operations except reload)
          @Required
        type: string
      file_user:
        description: |-
          File owner username
          @Description Username to set as file owner (required for chown operation)
        type: string
      link_path:
        description: |-
          Target path for symlink
          @Description Path for symlink target (required for symlink operation)
        type: string
      operation:
        description: |-
          Operation type: write, symlink, chown, reload, delete, read, check
          @Description Operation type to perform
          @Required
          @Enum(write,symlink,chown,reload,delete,read,check)
        type: string
    type: object
host: localhost:8256
info:
  contact: {}
  description: Path for symlink target (required for symlink operation)
  title: Alarm Rules Manager API
  version: "1.0"
paths:
...
  /api/v1/rules/{operation}:
    post:
      consumes:
      - application/json
      description: Handle various operations on rule files (write, symlink, chown,
        reload, delete, read, check)
      parameters:
      - description: Rule file operation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.RuleFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Operation successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Handle rule file operations
      tags:
      - rules
...
  /security_groups/{id}/rules:
    get:
      consumes:
      - application/json
      description: list secrules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.SecruleListResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/common.APIError'
      summary: list secrules
      tags:
      - Network
    post:
      consumes:
      - application/json
      description: create a secrule
      parameters:
      - description: Secrule create payload
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/apis.SecurityRulePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.SecruleResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.APIError'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/common.APIError'
      summary: create a secrule
      tags:
      - Network
  /security_groups/{id}/rules/{rule_id}:
    delete:
      consumes:
      - application/json
      description: delete a secrule
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.APIError'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/common.APIError'
      summary: delete a secrule
      tags:
      - Network
    get:
      consumes:
      - application/json
      description: get a secrule
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.SecruleResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.APIError'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/common.APIError'
      summary: get a secrule
      tags:
      - Network
...
swagger: "2.0"